requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
spacy
pandas
streamlit
reportlab
